requirements.txt
click
colorama
nbclean
pyyaml
black
